Fix remoteMediators for correct pagination of data. Add temporary solution with base logic for MovieListFragment.kt and MovieDetailFragment.kt. Update layouts.

+++ app/src/main/java/com/andreich/moviesearcher/domain/pojo/RatingDto.kt
@@ -6,6 +6,6 @@ data class RatingDto(
     @SerializedName("kp") var kp: Double? = null,
     @SerializedName("imdb") var imdb: Double? = null,
     @SerializedName("filmCritics") var filmCritics: Double? = null,
-    @SerializedName("russianFilmCritics") var russianFilmCritics: Int? = null,
+    @SerializedName("russianFilmCritics") var russianFilmCritics: Double? = null,
     @SerializedName("await") var await: String? = null
 )
